api/requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
PyYAML==6.0.2